[level 2] Title: 피보나치 수, Time: 0.33 ms, Memory: 4.15 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.cpp
@@ -1,11 +1,9 @@
 // https://school.programmers.co.kr/learn/courses/14743/lessons/116435
-#include <iostream>
-
 using namespace std;
-long long addDivideThem(int n, long long a=0, long long b=1){
+int addDivideThem(int n, int a=0, int b=1){
     if(n==0) return a;
     if(n==1) return b;
-    if(n==2) return (long long)(a+b)%1234567;
+    if(n==2) return (a+b)%1234567;
     if(n>2) return addDivideThem(n-1, b, (a+b)%1234567);
 }
